fixed sign up, sign in, sign out functions removed profile.js order of the screens: auth.js (for sign up and sign in) -> account.tsx (for update and sign out)
